+++ ext/polars/src/lib.rs
@@ -767,6 +767,7 @@ fn init(ruby: &Ruby) -> RbResult<()> {
         function!(functions::io::read_parquet_schema, 1),
     )?;
     class.define_singleton_method("collect_all", function!(functions::lazy::collect_all, 3))?;
+    class.define_singleton_method("explain_all", function!(functions::lazy::explain_all, 2))?;
     class.define_singleton_method(
         "collect_all_lazy",
         function!(functions::lazy::collect_all_lazy, 2),