+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - Added `explain_all` method to `Polars`
 - Fixed `storage_options` option for `scan_parquet` and `scan_ipc` methods
 - Fixed `Struct` construction with `nil` values
-- Fixed `read_database` for JSON columns with `NULL` values
+- Fixed `read_database` for JSON columns
 
 ## 0.25.0 (2026-02-18)